Move atomistics node package
